Move MultiFileReaderData to separate file
